be-ai/requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# Vector Database & Embeddings
chromadb==0.4.24
sentence-transformers==2.5.1
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu

# AI & LLM
# google-generativeai==0.4.1
python-dotenv==1.0.1
google-genai

# Utilities
pandas==2.2.1
requests==2.31.0